// src/app/component/thi-hanh-an/model/decidata.model.ts
export interface DeciDataModel {
    deci_data_code?: number;
    deciid?: string;
    decicode?: string;
    regicode?: string;
    crtDate?: Date | string;
    crtuser?: string;
    actdate?: Date | string;
    enforcetype?: string;
    agency_assigned?: string;
    time_reduction?: string;
    tyear?: string;
    tmonth?: string;
    tday?: string;
    fyear?: string;
    fmonth?: string;
    fday?: string;
    chell_year?: string;
    chell_month?: string;
    chell_day?: string;
}